fix: review fixes — consumer topics, int types, session query
- Add car-motion and final-classification to KafkaConsumerService
  topic subscription (critical: data was not reaching TimescaleDB)
- Change FinalClassificationEvent fields from byte/uint to int
  (matches project convention for all other event records)
- Fix PacketMapper to use (int) casts and cap NumCars at 20
- Fix session variable query in all 6 dashboards: DISTINCT with
  ORDER BY non-selected column is invalid PostgreSQL; use GROUP BY

+++ src/TelemetryIngester.Tests/Mapping/PacketMapperTests.cs
@@ -813,9 +813,9 @@ public sealed class PacketMapperTests
         Assert.Equal(3, @event.GridPosition);
         Assert.Equal(25, @event.Points);
         Assert.Equal(1, @event.NumPitStops);
-        Assert.Equal((byte)ResultStatus.Finished, @event.ResultStatus);
-        Assert.Equal((byte)ResultReason.Invalid, @event.ResultReason);
-        Assert.Equal(90500u, @event.BestLapTimeMs);
+        Assert.Equal((int)ResultStatus.Finished, @event.ResultStatus);
+        Assert.Equal((int)ResultReason.Invalid, @event.ResultReason);
+        Assert.Equal(90500, @event.BestLapTimeMs);
         Assert.Equal(5432.123, @event.TotalRaceTime);
         Assert.Equal(0, @event.PenaltiesTime);
         Assert.Equal(0, @event.NumPenalties);
